requirements.txt
torch>=2.1.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
matplotlib>=3.7.0
jinja2>=3.1.0
pytest>=7.0.0
